refactoring management commant, update README and requirements
