+++ README.md
@@ -3,10 +3,6 @@ api_yamdb
 
 надо не забыть заполнить
 
-python manage.py add_db users.csv user
-python manage.py add_db category.csv category
-python manage.py add_db genre.csv genre
-python manage.py add_db titles.csv title
-python manage.py add_db genre_title.csv genretitle
-python manage.py add_db review.csv review
-python manage.py add_db comments.csv comment
+для наполнения базы подготовленными данными необходимо выполнить эту команду:
+
+python manage.py import_data
