+++ api_yamdb/static/data/comments.csv
@@ -1,4 +1,4 @@
-id,review_id,text,author,pub_date
+id,review_id,text,author_id,pub_date
 1,6,"Ничего подобного, в фильме всё не так, и программирование тут вообще ни при чём!",102,2020-01-13T23:20:02.422Z
 2,6,"Ну надо же, не нашлось ничего лучшего, кроме как прокомментировать разговор про гамбургеры, будто в фильме ничего важнее этого нет",101,2020-01-13T23:20:02.422Z
 3,6,"Кстати, а что такое ""четверть фунта""? В граммах это сколько?",103,2020-01-13T23:20:02.422Z
