+++ api_yamdb/static/data/review.csv
@@ -1,4 +1,4 @@
-id,title_id,text,author,score,pub_date
+id,title_id,text,author_id,score,pub_date
 1,1,"Ставлю десять звёзд!
 ...Эти голоса были чище и светлее тех, о которых мечтали в этом сером, убогом месте. Как будто две птички влетели и своими голосами развеяли стены наших клеток, и на короткий миг каждый человек в Шоушенке почувствовал себя свободным.",100,10,2019-09-24T21:08:21.567Z
 2,1,"Не привыкай
