+++ api_yamdb/static/data/titles.csv
@@ -1,4 +1,4 @@
-id,name,year,category
+id,name,year,category_id
 1,Побег из Шоушенка,1994,1
 2,Крестный отец,1972,1
 3,12 разгневанных мужчин,1957,1
